move autogenerated code together into java module
