--- src/collections.rs
@@ -1,1 +1,0 @@
-pub mod list;
